Working fetchAPI function + basic barchart w/ retrieved data.
The fetchAPI function works both when called on page load (with useEffect) and onClick. There seems to be no useless rerenderings anymore. Data fetched is properly passed to drawChart() and the div is cleared on useEffect clean up to avoid chart multiplication. Not tested yet if the chart is properly updated when retrieving new data.