Updated GridChart to actually contain the real fetched data and display the doc names.
Duplicated ViewA into a ViewB in order to maintain the distinction between the two different data fetching.

Now data fetching in ViewB is agnostic wrt the actual data fetched, but the GridChart component is only rendered if data != null.

The document names are displayed with a temp hack, they are not actually contained in the rects.