Improved logic behind GridChart() setup: data is used directly to build it, and the chart's dimensions depend on the data.
Removed doc name from the squares in GridChart().
Unified createGridData() and updateGridData() functions, so that the actual info re each document is inizialized directly, instead than in a subsequent step.
createGridData() now returns an object containing the old gridData + the number of rows (totRanks) and columns (totRuns), which are necessary to rend the chart in D3.js.