Fixed the Big PoolDepth Debacle™.
All mentions of Pool Depth have been purged. In the backend, it has been completely removed from the API. In the frontend, it has been turned to runSize. It has been added as a prop for GridChart (and then for GridData), and it is used to decide the height of GridChart, by simply cutting off the json data received where it's not needed.

This gives better performance when changing the Run Size to be displayed, but received ALL data from the backend caused a bit of lag when chaning topics. For this reason, the maximum Run Size allowed has been put to 100, and its currently hard-coded in the backend, which only sends the first 100 elements for each run.