Restructured loading process + fixes.
Now the loaddata API returns two parameters: runsList and topicsList, which are used by the client to correctly display the charts and the Topic Picker.

The loading process was changed: the server no longer loads all runs + qrels on start up, but a call is made to the loaddata API by the client at first render.

Fixed a bug in the way the height of GridChart was computed.